+++ src/array/distinct/index.ts
@@ -2,17 +2,20 @@ import * as method from "./method";
 
 declare global {
   interface Array<T> {
-    distinct(): T[];
+    distinct(fn?: string | ((value: T, index: number, array: T[]) => any)): T[];
   }
 }
 
 /**
  * Returns all the distinct values of an array
  * @memberof Array
+ * @param {String|Function} [fn]
  * @returns {Array}
  * @example
  * [1, 2, 2, 3, 4, 4, 5].distinct(); // [1,2,3,4,5]
+ * [{foo:{bar:[0,1]}}, {foo:{bar:[4,1]}}].distinct("foo.bar[1]"); // [{foo:{bar:[0,1]}}]
+ * [1, 2, 2.2, 3, 4.9, 4, 5].distinct(Math.floor); // [1,2,3,4.9,5]
  */
-Array.prototype.distinct = function() {
-  return method(this);
+Array.prototype.distinct = function(fn) {
+  return method(this, fn);
 };