+++ src/number/methods.ts
@@ -1,9 +1,11 @@
 import * as digitize from "./digitize/method";
+import * as equals from "./equals/method";
 import * as inRange from "./inRange/method";
 import * as isNumber from "./isNumber/method";
 
 export {
   digitize,
+  equals,
   inRange,
   isNumber,
 };