+++ src/object/forEach/index.ts
@@ -1,15 +1,14 @@
 import * as method from "./method";
-import { addPrototype } from "../utils";
 
 declare global {
   interface Object {
-    $forEach(fn: (value: any, key: string | number, object: object) => void): void;
+    $forEach(fn: (value: any, key: string, object: object) => void): void;
   }
 }
 
 /**
  * Iterates the object by keys
- * @memberof Object.prototype
+ * @memberof Object
  * @param {Function} fn
  * @example
  * const users = {
@@ -18,8 +17,6 @@ declare global {
  * };
  * users.$forEach(u => console.log(u.age)); // console logs 40 and 1
  */
-function $forEach(this: object, fn: (value: any, key: string | number, object: object) => void) {
+Object.prototype.$forEach = function(fn) {
   return method(this, fn);
-}
-
-addPrototype("$forEach", $forEach);
+};