+++ src/object/forEach/method.ts
@@ -1,4 +1,12 @@
-const method = (obj: object, fn: (value: any, key: string | number, object: object) => void) =>
-  Object.keys(obj).forEach((key) => fn((obj as { [key: string]: any })[key], key, obj));
+const method = (obj: object, fn: (value: any, key: string, object: object) => void) => {
+  const keys = Object.keys(obj);
+  const length = keys.length;
+
+  for (let i = 0; i < length; i++) {
+    const key = keys[i];
+
+    fn((obj as { [key: string]: any })[key], key, obj);
+  }
+};
 
 export = method;