+++ src/object/map/method.ts
@@ -1,6 +1,6 @@
 import * as forEach from "../forEach/method";
 
-const method = (obj: object, fn: (value: any, key: string | number, object: object) => object): object => {
+const method = (obj: object, fn: (value: any, key: string, object: object) => { [key: string]: any }): object => {
   return Object.keys(obj).reduce((acc: { [key: string]: any }, k) => {
     forEach(fn((obj as { [key: string]: any })[k], k, obj), (value, key) => acc[key] = value);