+++ src/object/values/index.ts
@@ -1,21 +1,18 @@
 import * as method from "./method";
-import { addPrototype } from "../utils";
 
 declare global {
   interface Object {
-    $values(): number;
+    $values(): any[];
   }
 }
 
 /**
  * Get values of the object
- * @memberof Object.prototype
+ * @memberof Object
  * @returns {Object}
  * @example
- * { one: 1, two: 2, three: 3 }.$values(); // [1, 2, 3]
+ * ({ one: 1, two: 2, three: 3 }).$values(); // [1, 2, 3]
  */
-function $values(this: object): string[] {
+Object.prototype.$values = function() {
   return method(this);
-}
-
-addPrototype("$values", $values);
+};