+++ src/string/reduce/method.ts
@@ -1,0 +1,7 @@
+const method = (
+  str: string,
+  fn: (prev: any, char: string, index: number, chars: string[]) => any,
+  initialValue: any = undefined,
+) => str.split("").reduce(fn, initialValue);
+
+export = method;